+++ packages/picasso.js/src/core/chart-components/box/__tests__/box-shapes.spec.js
@@ -1,4 +1,5 @@
 import {
+  oob,
   box,
   verticalLine,
   horizontalLine,
@@ -6,6 +7,53 @@ import {
 } from '../box-shapes';
 
 describe('box shapes', () => {
+  describe('out of bounds arrows', () => {
+    it('should create a correct out of bounds shape at the bottom of the graph', () => {
+      const item = {
+        major: 0.3,
+        oob: {
+          size: 10
+        }
+      };
+
+      // We mock the symbol() function here
+      let result = oob({
+        value: 1, item, boxCenter: item.major, rendWidth: 100, rendHeight: 100, flipXY: false, symbol: v => v
+      });
+
+      expect(result).to.eql({
+        size: 10,
+        startAngle: -90,
+        x: 30,
+        y: 95
+      });
+
+      item.major = 0.5;
+    });
+
+    it('should create a correct out of bounds shape at the top of the graph', () => {
+      const item = {
+        major: 0.75,
+        oob: {
+          size: 10
+        }
+      };
+
+      // We mock the symbol() function here
+      let result = oob({
+        value: 0, item, boxCenter: item.major, rendWidth: 100, rendHeight: 100, flipXY: false, symbol: v => v
+      });
+
+      expect(result).to.eql({
+        size: 10,
+        startAngle: 90,
+        x: 75,
+        y: 5
+      });
+    });
+  });
+
+
   describe('box', () => {
     it('should create a correct box shape', () => {
       const item = {
@@ -21,7 +69,7 @@ describe('box shapes', () => {
       const boxPadding = 0;
 
       let result = box({
-        item, boxWidth, boxPadding, rendwidth: 100, rendheight: 100, flipXY: false
+        item, boxWidth, boxPadding, rendWidth: 100, rendHeight: 100, flipXY: false
       });
 
       expect(result).to.eql({
@@ -39,7 +87,7 @@ describe('box shapes', () => {
 
 
       result = box({
-        item, boxWidth, boxPadding, rendwidth: 200, rendheight: 100, flipXY: false
+        item, boxWidth, boxPadding, rendWidth: 200, rendHeight: 100, flipXY: false
       });
 
       expect(result).to.eql({
@@ -56,7 +104,7 @@ describe('box shapes', () => {
       });
 
       result = box({
-        item, boxWidth, boxPadding, rendwidth: 100, rendheight: 100, flipXY: true
+        item, boxWidth, boxPadding, rendWidth: 100, rendHeight: 100, flipXY: true
       });
 
       expect(result).to.eql({
@@ -74,7 +122,7 @@ describe('box shapes', () => {
 
       item.data = { test: true };
       result = box({
-        item, boxWidth, boxPadding, rendwidth: 200, rendheight: 100, flipXY: true
+        item, boxWidth, boxPadding, rendWidth: 200, rendHeight: 100, flipXY: true
       });
 
       expect(result).to.eql({
@@ -103,7 +151,7 @@ describe('box shapes', () => {
       };
 
       let result = verticalLine({
-        item, from: 0.25, to: 0.75, boxCenter: 0.5, rendwidth: 100, rendheight: 100, flipXY: false
+        item, from: 0.25, to: 0.75, boxCenter: 0.5, rendWidth: 100, rendHeight: 100, flipXY: false
       });
 
       expect(result).to.eql({
@@ -120,7 +168,7 @@ describe('box shapes', () => {
       });
 
       result = verticalLine({
-        item, from: 0.25, to: 0.75, boxCenter: 0.5, rendwidth: 200, rendheight: 100, flipXY: false
+        item, from: 0.25, to: 0.75, boxCenter: 0.5, rendWidth: 200, rendHeight: 100, flipXY: false
       });
 
       expect(result).to.eql({
@@ -138,7 +186,7 @@ describe('box shapes', () => {
 
       item.data = { id: 3 };
       result = verticalLine({
-        item, from: 0.25, to: 0.75, boxCenter: 0.5, rendwidth: 200, rendheight: 100, flipXY: true
+        item, from: 0.25, to: 0.75, boxCenter: 0.5, rendWidth: 200, rendHeight: 100, flipXY: true
       });
 
       expect(result).to.eql({
@@ -170,7 +218,7 @@ describe('box shapes', () => {
       };
 
       let result = horizontalLine({
-        item, key: 'med', position: 0.5, width: 0.5, boxCenter: 0.5, rendwidth: 100, rendheight: 100, flipXY: false
+        item, key: 'med', position: 0.5, width: 0.5, boxCenter: 0.5, rendWidth: 100, rendHeight: 100, flipXY: false
       });
 
       expect(result).to.eql({
@@ -191,7 +239,7 @@ describe('box shapes', () => {
       });
 
       result = horizontalLine({
-        item, key: 'low', position: 0.5, width: 0.5, boxCenter: 0.5, rendwidth: 200, rendheight: 100, flipXY: false
+        item, key: 'low', position: 0.5, width: 0.5, boxCenter: 0.5, rendWidth: 200, rendHeight: 100, flipXY: false
       });
 
       expect(result).to.eql({
@@ -213,7 +261,7 @@ describe('box shapes', () => {
 
       item.data = { origin: 'test' };
       result = horizontalLine({
-        item, key: 'med', position: 0.5, width: 0.5, boxCenter: 0.5, rendwidth: 200, rendheight: 100, flipXY: true
+        item, key: 'med', position: 0.5, width: 0.5, boxCenter: 0.5, rendWidth: 200, rendHeight: 100, flipXY: true
       });
 
       expect(result).to.eql({