+++ packages/picasso.js/src/core/symbols/bar.js
@@ -29,7 +29,6 @@ function bar(options) {
   const rect = pointsToRect(points);
   rect.type = 'rect';
   rect.fill = 'black';
-  rect.collider = { type: 'rect' };
 
   return rect;
 }