+++ packages/picasso.js/src/core/symbols/line.js
@@ -1,3 +1,14 @@
+function createRectCollider({ x, y, size }) {
+  const r = size / 2;
+  return {
+    type: 'rect',
+    x: x - r,
+    y: y - r,
+    width: size,
+    height: size
+  };
+}
+
 /**
  * @extends symbol-config
  * @typedef {object} symbol--line
@@ -16,7 +27,8 @@ function line(options) {
     x1: x - (isVertical ? 0 : r),
     y1: y - (isVertical ? r : 0),
     x2: x + (isVertical ? 0 : r),
-    y2: y + (isVertical ? r : 0)
+    y2: y + (isVertical ? r : 0),
+    collider: createRectCollider(options) // TODO Use visual collider?
   };
 }