+++ packages/picasso.js/src/core/utils/class-string.js
@@ -1,5 +1,6 @@
 /**
  * Concatenate object keys into a space separated string. Use for transforming a 'class name map' into a class string
+ * @private
  * @param  {Object} classMap Object with class names as keys and true or false values depending on if they should be in the returned class string or not
  * @return {String} Space separated string with class names
  */