+++ packages/picasso.js/src/web/components/brush-range/brush-area-dir.js
@@ -174,7 +174,9 @@ const brushAreaDirectionalComponent = {
     this.rect = {
       x: 0, y: 0, width: 0, height: 0
     };
-    this.state = {};
+    this.state = {
+      key: this.settings.key || 'brush-area-dir'
+    };
   },
   beforeRender(opts) {
     this.rect = opts.size;